fix: visualizador nao consegue editar etapas de planning

+++ app/Livewire/Planning/DataExtraction/Question.php
@@ -9,15 +9,20 @@ use App\Models\Project\Planning\DataExtraction\Question as QuestionModel;
 use App\Models\Project\Planning\DataExtraction\QuestionTypes as QuestionTypesModel;
 use App\Utils\ActivityLogHelper as Log;
 use App\Utils\ToastHelper;
+use App\Traits\ProjectPermissions;
 
 class Question extends Component
 {
+
+    use ProjectPermissions;
+
     public $currentProject;
     public $currentQuestion;
     public $questions = [];
     public $questionId;
     public $type;
     public $questionTypes = [];
+    private $toastMessages = 'project/planning.data-extraction.toasts';
 
     /**
      * Fields to be filled by the form.
@@ -83,6 +88,11 @@ class Question extends Component
      */
     public function updateQuestions()
     {
+
+        if (!$this->checkEditPermission($this->toastMessages . '.denied')) {
+            return;
+        }
+
         $this->questions = QuestionModel::where(
             'id_project',
             $this->currentProject->id_project
@@ -104,6 +114,11 @@ class Question extends Component
      */
     public function submit()
     {
+
+        if (!$this->checkEditPermission($this->toastMessages . '.denied')) {
+            return;
+        }
+
         $this->validate();
 
         // Verifica se o 'id' da questão já existe no projeto atual (em modo de criação)
@@ -173,6 +188,10 @@ class Question extends Component
     #[On('data-extraction-table-edit-question')]
     public function edit(string $questionId)
     {
+        if (!$this->checkEditPermission($this->toastMessages . '.denied')) {
+            return;
+        }
+
         $this->currentQuestion = QuestionModel::where('id_project', $this->currentProject->id_project)
             ->where('id_de', $questionId)
             ->first();
@@ -212,6 +231,11 @@ class Question extends Component
     #[On('data-extraction-table-delete-question')]
     public function delete(string $questionId)
     {
+
+        if (!$this->checkEditPermission($this->toastMessages . '.denied')) {
+            return;
+        }
+
         try {
             $currentQuestion = QuestionModel::where('id_project', $this->currentProject->id_project)->where('id_de', $questionId)->first();
             $currentQuestion->delete();
